Fix comment-2
