+++ paimon-core/src/main/java/org/apache/paimon/table/source/AbstractDataTableRead.java
@@ -18,22 +18,14 @@
 
 package org.apache.paimon.table.source;
 
-import org.apache.paimon.CoreOptions;
-import org.apache.paimon.catalog.CatalogContext;
 import org.apache.paimon.catalog.TableQueryAuthResult;
-import org.apache.paimon.data.Blob;
-import org.apache.paimon.data.BlobView;
-import org.apache.paimon.data.BlobViewResolver;
-import org.apache.paimon.data.BlobViewStruct;
 import org.apache.paimon.data.InternalRow;
 import org.apache.paimon.disk.IOManager;
 import org.apache.paimon.predicate.Predicate;
 import org.apache.paimon.predicate.PredicateProjectionConverter;
 import org.apache.paimon.reader.RecordReader;
 import org.apache.paimon.schema.TableSchema;
-import org.apache.paimon.types.DataTypeRoot;
 import org.apache.paimon.types.RowType;
-import org.apache.paimon.utils.BlobViewLookup;
 import org.apache.paimon.utils.ListUtils;
 import org.apache.paimon.utils.ProjectedRow;
 
@@ -42,11 +34,9 @@ import javax.annotation.Nullable;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashSet;
-import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
-import java.util.function.Supplier;
 
 import static org.apache.paimon.predicate.PredicateVisitor.collectFieldNames;
 
@@ -57,20 +47,9 @@ public abstract class AbstractDataTableRead implements InnerTableRead {
     private boolean executeFilter = false;
     private Predicate predicate;
     private final TableSchema schema;
-    @Nullable private final CatalogContext catalogContext;
-    @Nullable private final Supplier<InnerTableRead> readFactory;
 
     public AbstractDataTableRead(TableSchema schema) {
-        this(schema, null, null);
-    }
-
-    public AbstractDataTableRead(
-            TableSchema schema,
-            @Nullable CatalogContext catalogContext,
-            @Nullable Supplier<InnerTableRead> readFactory) {
         this.schema = schema;
-        this.catalogContext = catalogContext;
-        this.readFactory = readFactory;
     }
 
     public abstract void applyReadType(RowType readType);
@@ -113,27 +92,35 @@ public abstract class AbstractDataTableRead implements InnerTableRead {
 
     protected void configurePrescanRead(InnerTableRead prescanRead) {}
 
+    protected TableSchema schema() {
+        return schema;
+    }
+
+    protected RowType currentReadType() {
+        return readType == null ? schema.logicalRowType() : readType;
+    }
+
+    @Nullable
+    protected Predicate predicate() {
+        return predicate;
+    }
+
     @Override
-    public final RecordReader<InternalRow> createReader(Split split) throws IOException {
-        TableQueryAuthResult authResult = null;
+    public RecordReader<InternalRow> createReader(Split split) throws IOException {
+        QueryAuthContext queryAuthContext = unwrapQueryAuthSplit(split);
+        return createDataReader(queryAuthContext.split(), queryAuthContext.authResult());
+    }
+
+    protected final QueryAuthContext unwrapQueryAuthSplit(Split split) {
         if (split instanceof QueryAuthSplit) {
             QueryAuthSplit authSplit = (QueryAuthSplit) split;
-            split = authSplit.split();
-            authResult = authSplit.authResult();
+            return new QueryAuthContext(authSplit.split(), authSplit.authResult());
         }
+        return new QueryAuthContext(split, null);
+    }
 
-        if (catalogContext != null) {
-            RowType rowType = this.readType == null ? schema.logicalRowType() : this.readType;
-            int[] blobViewFields = blobViewFields(rowType);
-            if (blobViewFields.length > 0) {
-                if (readFactory == null) {
-                    throw new IllegalStateException(
-                            "Cannot read blob-view-field fields without a readFactory.");
-                }
-                return createBlobViewReader(split, authResult, blobViewFields);
-            }
-        }
-
+    protected final RecordReader<InternalRow> createDataReader(
+            Split split, @Nullable TableQueryAuthResult authResult) throws IOException {
         RecordReader<InternalRow> reader;
         if (authResult == null) {
             reader = reader(split);
@@ -145,70 +132,6 @@ public abstract class AbstractDataTableRead implements InnerTableRead {
         }
 
         return reader;
-    }
-
-    private int[] blobViewFields(RowType rowType) {
-        Set<String> blobViewFieldNames = CoreOptions.fromMap(schema.options()).blobViewField();
-        if (blobViewFieldNames.isEmpty()) {
-            return new int[0];
-        }
-
-        return rowType.getFields().stream()
-                .filter(
-                        field ->
-                                field.type().is(DataTypeRoot.BLOB)
-                                        && blobViewFieldNames.contains(field.name()))
-                .mapToInt(field -> rowType.getFieldIndex(field.name()))
-                .toArray();
-    }
-
-    private RecordReader<InternalRow> createBlobViewReader(
-            Split split, @Nullable TableQueryAuthResult authResult, int[] blobViewFields)
-            throws IOException {
-        RowType rowType = this.readType == null ? schema.logicalRowType() : this.readType;
-        RowType blobViewOnlyType = rowType.project(blobViewFields);
-        InnerTableRead prescanRead = readFactory.get();
-        prescanRead.withReadType(blobViewOnlyType);
-        if (predicate != null) {
-            prescanRead.withFilter(predicate);
-        }
-        configurePrescanRead(prescanRead);
-        Split prescanSplit = authResult != null ? new QueryAuthSplit(split, authResult) : split;
-        LinkedHashSet<BlobViewStruct> viewStructs = new LinkedHashSet<>();
-        RecordReader<InternalRow> prescanReader = prescanRead.createReader(prescanSplit);
-        try {
-            prescanReader.forEachRemaining(
-                    row -> {
-                        for (int i = 0; i < blobViewFields.length; i++) {
-                            if (row.isNullAt(i)) {
-                                continue;
-                            }
-                            Blob blob = row.getBlob(i);
-                            if (!(blob instanceof BlobView)) {
-                                throw new IllegalArgumentException(
-                                        "blob-view-field requires blob field value to be a "
-                                                + "serialized BlobViewStruct.");
-                            }
-                            viewStructs.add(((BlobView) blob).viewStruct());
-                        }
-                    });
-        } finally {
-            prescanReader.close();
-        }
-
-        BlobViewResolver resolver =
-                BlobViewLookup.createResolver(catalogContext, new ArrayList<>(viewStructs));
-
-        RecordReader<InternalRow> reader =
-                authResult == null ? reader(split) : authedReader(split, authResult);
-        if (executeFilter) {
-            reader = executeFilter(reader);
-        }
-        Set<Integer> blobViewFieldSet = new HashSet<>();
-        for (int field : blobViewFields) {
-            blobViewFieldSet.add(field);
-        }
-        return reader.transform(row -> new BlobViewResolvingRow(row, blobViewFieldSet, resolver));
     }
 
     private RecordReader<InternalRow> authedReader(Split split, TableQueryAuthResult authResult)
@@ -259,5 +182,25 @@ public abstract class AbstractDataTableRead implements InnerTableRead {
 
         Predicate finalFilter = predicate;
         return reader.filter(finalFilter::test);
+    }
+
+    protected static class QueryAuthContext {
+
+        private final Split split;
+        @Nullable private final TableQueryAuthResult authResult;
+
+        private QueryAuthContext(Split split, @Nullable TableQueryAuthResult authResult) {
+            this.split = split;
+            this.authResult = authResult;
+        }
+
+        protected Split split() {
+            return split;
+        }
+
+        @Nullable
+        protected TableQueryAuthResult authResult() {
+            return authResult;
+        }
     }
 }
