+++ paimon-core/src/main/java/org/apache/paimon/table/source/AppendTableRead.java
@@ -18,7 +18,6 @@
 
 package org.apache.paimon.table.source;
 
-import org.apache.paimon.catalog.CatalogContext;
 import org.apache.paimon.data.InternalRow;
 import org.apache.paimon.operation.MergeFileSplitRead;
 import org.apache.paimon.operation.SplitRead;
@@ -36,13 +35,12 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Function;
-import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
 /**
  * An abstraction layer above {@link MergeFileSplitRead} to provide reading of {@link InternalRow}.
  */
-public final class AppendTableRead extends AbstractDataTableRead {
+public class AppendTableRead extends AbstractDataTableRead {
 
     private final List<SplitReadProvider> readProviders;
 
@@ -54,15 +52,7 @@ public final class AppendTableRead extends AbstractDataTableRead {
     public AppendTableRead(
             List<Function<SplitReadConfig, SplitReadProvider>> providerFactories,
             TableSchema schema) {
-        this(providerFactories, schema, null, null);
-    }
-
-    public AppendTableRead(
-            List<Function<SplitReadConfig, SplitReadProvider>> providerFactories,
-            TableSchema schema,
-            @Nullable CatalogContext catalogContext,
-            @Nullable Supplier<InnerTableRead> readFactory) {
-        super(schema, catalogContext, readFactory);
+        super(schema);
         this.readProviders =
                 providerFactories.stream()
                         .map(factory -> factory.apply(this::config))
